refactoring deposits query, refactoring fillable in models(add direction)

+++ database/migrations/2024_01_08_185118_create_deposits_table.php
@@ -23,8 +23,8 @@ return new class extends Migration
             $table->foreignIdFor(Bank::class);
 
             $table->string('type')->default(DepositTypeEnum::deposit()->value);
-            $table->float('price');
-            $table->float('percent');
+            $table->double('price');
+            $table->double('percent');
             $table->dateTime('expiration_date')->nullable();
